chore: use unified store with slice to better contorl global state
