+++ packages/toolkit/src/lib/index.ts
@@ -9,6 +9,7 @@ export * from "./store";
 export * from "./table";
 export * from "./type";
 export * from "./use-instill-form";
+export * from "./use-instill-store";
 export * from "./utility";
 export * from "./vdp-sdk";
 export * from "./useResourceAdditionalForm";
