+++ packages/toolkit/src/lib/use-instill-store/smartHintSlice.ts
@@ -1,0 +1,12 @@
+import { StateCreator } from "zustand";
+import { InstillStore, InstillStoreMutators, SmartHintSlice } from "./types";
+
+export const createSmartHintSlice: StateCreator<
+  InstillStore,
+  InstillStoreMutators,
+  [],
+  SmartHintSlice
+> = (set) => ({
+  smartHints: [],
+  updateSmartHints: (smartHints) => set({ smartHints }),
+});
