+++ packages/toolkit/src/lib/use-smart-hint/types.ts
@@ -1,0 +1,4 @@
+export type SmartHint = {
+  path: string;
+  instillFormat: string;
+};
