+++ packages/toolkit/src/lib/use-smart-hint/useSmartHint.ts
@@ -1,0 +1,9 @@
+import { useEffect } from "react";
+import { Pipeline } from "../vdp-sdk";
+import { Nullable } from "../type";
+
+export const useSmartHint = (pipeline: Nullable<Pipeline>) => {
+  useEffect(() => {
+    console.log("placeholder");
+  }, [pipeline]);
+};
