+++ packages/toolkit/src/view/data/DataComponentAutoForm.tsx
@@ -1,14 +1,18 @@
-import { ConnectorDefinition, GeneralRecord } from "../../lib";
+import { useShallow } from "zustand/react/shallow";
+
 import {
-  PipelineBuilderStore,
+  ConnectorDefinition,
+  GeneralRecord,
+  InstillStore,
+  useInstillStore,
+} from "../../lib";
+import {
   PipelineComponentReference,
   composeEdgesFromReferences,
   extractReferencesFromConfiguration,
   recursiveReplaceNullAndEmptyStringWithUndefined,
   recursiveTransformToString,
-  usePipelineBuilderStore,
 } from "../pipeline-builder";
-import { shallow } from "zustand/shallow";
 import { ResourceComponentForm } from "../resource";
 
 export type DataComponentAutoFormProps = {
@@ -17,13 +21,13 @@ export type DataComponentAutoFormProps = {
   disabledAll?: boolean;
 };
 
-const pipelineBuilderSelector = (state: PipelineBuilderStore) => ({
-  nodes: state.nodes,
-  updateNodes: state.updateNodes,
-  updateEdges: state.updateEdges,
-  selectedConnectorNodeId: state.selectedConnectorNodeId,
-  updateSelectedConnectorNodeId: state.updateSelectedConnectorNodeId,
-  updatePipelineRecipeIsDirty: state.updatePipelineRecipeIsDirty,
+const selector = (store: InstillStore) => ({
+  nodes: store.nodes,
+  updateNodes: store.updateNodes,
+  updateEdges: store.updateEdges,
+  selectedConnectorNodeId: store.selectedConnectorNodeId,
+  updateSelectedConnectorNodeId: store.updateSelectedConnectorNodeId,
+  updatePipelineRecipeIsDirty: store.updatePipelineRecipeIsDirty,
 });
 
 export const DataComponentAutoForm = (props: DataComponentAutoFormProps) => {
@@ -34,7 +38,7 @@ export const DataComponentAutoForm = (props: DataComponentAutoFormProps) => {
     selectedConnectorNodeId,
     updateSelectedConnectorNodeId,
     updatePipelineRecipeIsDirty,
-  } = usePipelineBuilderStore(pipelineBuilderSelector, shallow);
+  } = useInstillStore(useShallow(selector));
 
   /* eslint-disable-next-line  @typescript-eslint/no-explicit-any */
   function onSubmit(data: any) {
