+++ packages/toolkit/src/view/pipeline-builder/index.ts
@@ -6,4 +6,3 @@ export * from "./Flow";
 export * from "./FlowControl";
 export * from "./RightPanel";
 export * from "./type";
-export * from "./usePipelineBuilderStore";
