+++ packages/toolkit/src/view/pipeline/index.ts
@@ -2,4 +2,3 @@ export * from "./PipelineBuilderMainView";
 export * from "./PipelineListPageMainView";
 export * from "./PipelineTablePlaceholder";
 export * from "./PipelinesTable";
-export * from "./PipelineTable";
